win32: fix readline import for python >=3.13
In many parts of the code, including in the system library cmd, the missing readline library is imported.

To prevent an import error when import readline is executed in such places, we assign:

sys.modules["readline"] = _pyrepl.readline

This prevents an import error when import readline is executed in such places.

related: https://github.com/bretello/pdbpp/issues/47

This fixes the issue described in the KeyError with pyreadline3 on windows